indexer/pyproject.toml
[project]
name = "wiki-rag-indexer"
version = "0.2.0"
requires-python = ">=3.11"
dependencies = [
    "azure-storage-blob>=12.24",
    "llama-index>=0.10",
    "llama-index-vector-stores-postgres>=0.1",
    "llama-index-embeddings-openai>=0.1",
    "psycopg2-binary>=2.9",
    "requests>=2.32",
    "openai>=1.0"
]
